split main.go into multiple packages
