+++ pkg/conversion/convert.go
@@ -1,0 +1,496 @@
+package conversion
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"reflect"
+	"strconv"
+	"strings"
+
+	"github.com/fabianvf/windup-rulesets-yaml/pkg/windup"
+	"gopkg.in/yaml.v2"
+)
+
+func ConvertWindupRulesetsToAnalyzer(windups []windup.Ruleset, baseLocation, outputDir string) (map[string][][]map[string]interface{}, error) {
+	outputRulesets := map[string][][]map[string]interface{}{}
+	for _, windupRuleset := range windups {
+		ruleset := ConvertWindupRulesetToAnalyzer(windupRuleset)
+		yamlPath := strings.Replace(filepath.Join(outputDir, strings.Replace(windupRuleset.SourceFile, baseLocation, "", 1)), ".xml", ".yaml", 1)
+		if reflect.DeepEqual(ruleset, map[string]interface{}{}) {
+			continue
+		}
+		outputRulesets[yamlPath] = append(outputRulesets[yamlPath], ruleset)
+	}
+	for path, ruleset := range outputRulesets {
+		dirName := filepath.Dir(path)
+		err := os.MkdirAll(dirName, 0777)
+		if err != nil {
+			fmt.Printf("Skipping because of an error creating %s: %s\n", path, err.Error())
+			continue
+		}
+		err = writeYAML(ruleset, path)
+		if err != nil {
+			fmt.Printf("Skipping because of an error writing to %s: %s\n", path, err.Error())
+			continue
+		}
+	}
+	return outputRulesets, nil
+}
+
+func ConvertWindupRulesetToAnalyzer(ruleset windup.Ruleset) []map[string]interface{} {
+	// TODO Ruleset.Metadata
+	// TODO Ruleset.PackageMapping
+	// TODO Ruleset.Filemapping
+	// TODO Ruleset.Javaclassignore
+	rules := []map[string]interface{}{}
+	for i, windupRule := range ruleset.Rules.Rule {
+		// formatted, _ := yaml.Marshal(windupRule)
+		// fmt.Println(string(formatted))
+		rule := map[string]interface{}{
+			"ruleID": ruleset.SourceFile + "-" + strconv.Itoa(i),
+		}
+		where := flattenWhere(windupRule.Where)
+		if !reflect.DeepEqual(windupRule.When, windup.When{}) {
+			when := convertWindupWhenToAnalyzer(windupRule.When, where)
+			if len(when) == 1 {
+				rule["when"] = when[0]
+			} else if len(when) > 1 {
+				rule["when"] = map[string]interface{}{"or": when}
+			} else {
+				continue
+			}
+		} else {
+			continue
+		}
+
+		// TODO Rule.Perform
+		if !reflect.DeepEqual(windupRule.Perform, windup.Iteration{}) {
+			perform := convertWindupPerformToAnalyzer(windupRule.Perform, where)
+			// TODO only support a single action in perform right now, default to hint
+			tags, ok := perform["tag"].([]string)
+			if !ok {
+				tags = nil
+			}
+			if perform["message"] != nil {
+				rule["message"] = perform["message"]
+			} else if len(tags) != 0 {
+				rule["tag"] = tags
+			} else {
+				fmt.Println("\n\nNo action parsed\n\n")
+				continue
+			}
+			// for k, v := range perform {
+			// 	rule[k] = v
+			// }
+		} else {
+			continue
+		}
+
+		// TODO - Iteration
+		// TODO Rule.Otherwise
+		rules = append(rules, rule)
+	}
+	return rules
+}
+
+func convertWindupWhenToAnalyzer(windupWhen windup.When, where map[string]string) []map[string]interface{} {
+	//
+	// TODO Rule.When
+	// TODO - Graphquery
+	conditions := []map[string]interface{}{}
+	if windupWhen.Graphquery != nil {
+		for _, gq := range windupWhen.Graphquery {
+			// JarArchiveModel is a special case for deps
+			// that are actually included in a lib folder as jars
+			if gq.Discriminator == "JarArchiveModel" {
+				conditions = append(conditions, map[string]interface{}{"java.dependency": convertWindupGraphQueryJarArchiveModel(gq)})
+			} else if gq.Discriminator == "TechnologyTagModel" {
+				conditions = append(conditions, map[string]interface{}{"builtin.hasTags": []string{gq.Property.Value}})
+			}
+		}
+	}
+	if windupWhen.Project != nil {
+		for _, pc := range windupWhen.Project {
+			conditions = append(conditions, map[string]interface{}{"java.dependency": convertWindupDependencyToAnalyzer(pc.Artifact)})
+		}
+	}
+	if windupWhen.Dependency != nil {
+		for _, dc := range windupWhen.Dependency {
+			conditions = append(conditions, map[string]interface{}{"java.dependency": convertWindupDependencyToAnalyzer(dc)})
+		}
+	}
+	if windupWhen.And != nil {
+		whens := []map[string]interface{}{}
+		for _, condition := range windupWhen.And {
+			converted := convertWindupWhenToAnalyzer(condition, where)
+			for _, c := range converted {
+				whens = append(whens, c)
+			}
+		}
+		conditions = append(conditions, map[string]interface{}{"and": whens})
+	}
+	if windupWhen.Or != nil {
+		whens := []map[string]interface{}{}
+		for _, condition := range windupWhen.Or {
+			converted := convertWindupWhenToAnalyzer(condition, where)
+			for _, c := range converted {
+				whens = append(whens, c)
+			}
+		}
+		conditions = append(conditions, map[string]interface{}{"or": whens})
+	}
+	if windupWhen.Not != nil {
+		for _, condition := range windupWhen.Not {
+			converted := convertWindupWhenToAnalyzer(condition, where)
+			for _, c := range converted {
+				c["not"] = true
+				conditions = append(conditions, c)
+			}
+		}
+	}
+	if windupWhen.Filecontent != nil {
+		for _, fc := range windupWhen.Filecontent {
+			condition := map[string]interface{}{
+				"builtin.filecontent": strings.Replace(substituteWhere(where, fc.Pattern), "{*}", "*", -1),
+			}
+			if fc.As != "" {
+				condition["as"] = fc.As
+			}
+			if fc.From != "" {
+				condition["from"] = fc.From
+			}
+			conditions = append(conditions, condition)
+		}
+	}
+	if windupWhen.Javaclass != nil {
+		for _, jc := range windupWhen.Javaclass {
+			pattern := strings.Replace(substituteWhere(where, jc.References), "{*}", "*", -1)
+			pattern = strings.Replace(pattern, "(*)", "*", -1)
+			pattern = strings.Replace(pattern, ".*", "*", -1)
+			// Make some .* regesx's more generic.
+			// TODO: we may need to come back and figure out if we have to deal with
+			// capture groups and stuff.
+			pattern = strings.Replace(pattern, `(\.*)?\.`, ".*", -1)
+			pattern = strings.Replace(pattern, `.[^.]+`, "*", -1)
+			pattern = strings.Replace(pattern, `[^.]+`, "*", -1)
+			pattern = strings.Replace(pattern, `(\.[^.]*)*`, ".*", -1)
+			pattern = strings.Replace(pattern, `(?+[^.]*`, "", -1)
+			if jc.Location != nil {
+				for _, location := range jc.Location {
+					condition := map[string]interface{}{
+						"java.referenced": map[string]interface{}{
+							"location": location,
+							"pattern":  pattern,
+							// TODO handle jc.Annotationtype
+							// TODO handle jc.Annotationlist
+							// TODO handle jc.Annotationliteral
+							// TODO handle jc.MatchesSource
+							// TODO handle jc.In
+						},
+					}
+					if jc.As != "" {
+						condition["as"] = jc.As
+						// TODO (shurley): Only set when something is going to use this as block
+						// condition["ignore"] = true
+					}
+					if jc.From != "" {
+						condition["from"] = jc.From
+					}
+					conditions = append(conditions, condition)
+				}
+			} else {
+				condition := map[string]interface{}{
+					"java.referenced": map[string]interface{}{
+						"pattern": pattern,
+						// TODO handle jc.Annotationtype
+						// TODO handle jc.Annotationlist
+						// TODO handle jc.Annotationliteral
+						// TODO handle jc.MatchesSource
+						// TODO handle jc.In
+					},
+				}
+				if jc.As != "" {
+					condition["as"] = jc.As
+					// TODO (shurley): Only set when something is going to use this as block
+					// condition["ignore"] = true
+				}
+				if jc.From != "" {
+					condition["from"] = jc.From
+				}
+				conditions = append(conditions, condition)
+			}
+		}
+	}
+	if windupWhen.Xmlfile != nil {
+		for _, xf := range windupWhen.Xmlfile {
+			if xf.Matches == "" {
+				// TODO handle systemid and publicid
+				continue
+			}
+			// TODO find an actual way to deal with namespaces
+			matches := xf.Matches
+			if xf.Namespace != nil {
+				for _, ns := range xf.Namespace {
+					matches = strings.Replace(strings.Replace(matches, "windup:", "", -1), ns.Prefix+":", "", -1)
+				}
+			}
+			xmlCond := map[string]interface{}{
+				"xpath": substituteWhere(where, matches),
+			}
+			// TODO We don't support regexes here, may need to break it out into a separate lookup that gets passed through
+			if xf.In != "" {
+				in := substituteWhere(where, xf.In)
+				if strings.Contains(in, "{*}") {
+					conditions = append(conditions, map[string]interface{}{
+						"builtin.file": strings.Replace(xf.In, "{*}", "*", -1),
+						"as":           "xmlfiles",
+						"ignore":       true,
+					})
+					xmlCond["from"] = "xmlfiles"
+					xmlCond["filepaths"] = "{{xmlfiles.filepaths}}"
+				} else if strings.Contains(in, "|") {
+					filepaths := []string{}
+					pieces := strings.Split(in, "|")
+					for _, piece := range pieces {
+						filepaths = append(filepaths, strings.Trim(piece, "()"))
+					}
+					xmlCond["filepaths"] = pieces
+				} else {
+					xmlCond["filepaths"] = []string{in}
+				}
+			}
+			condition := map[string]interface{}{
+				"builtin.xml": xmlCond,
+			}
+			if xf.As != "" {
+				condition["as"] = xf.As
+				// TODO this is probably a dumb assumption
+				//	condition["ignore"] = true
+			}
+			// TODO What should we do if there's already a from because we had to split the task?
+			if xf.From != "" {
+				condition["from"] = xf.From
+			}
+			conditions = append(conditions, condition)
+		}
+	}
+
+	if windupWhen.File != nil {
+		for _, f := range windupWhen.File {
+			condition := map[string]interface{}{
+				"builtin.file": strings.Replace(substituteWhere(where, f.Filename), "{*}", "*", -1),
+			}
+			if f.As != "" {
+				condition["as"] = f.As
+				// TODO this is probably a dumb assumption
+				//condition["ignore"] = true
+			}
+			if f.From != "" {
+				condition["from"] = f.From
+			}
+			conditions = append(conditions, condition)
+		}
+	}
+	if windupWhen.Fileexists != nil {
+		for _, f := range windupWhen.Fileexists {
+			condition := map[string]interface{}{
+				"builtin.file": strings.Replace(substituteWhere(where, f.Filename), "{*}", "*", -1),
+			}
+			conditions = append(conditions, condition)
+		}
+	}
+	// TODO below here
+
+	// What is this ???
+	if windupWhen.True != "" {
+		// conditions = append(conditions, map[string]interface{}{"true": nil})
+	}
+	// What is this ???
+	if windupWhen.False != "" {
+		// conditions = append(conditions, map[string]interface{}{"false": nil})
+	}
+	// What is this ???
+	if windupWhen.Iterablefilter != nil {
+		// conditions = append(conditions, map[string]interface{}{"iterable-filter": nil})
+	}
+	// What is this ???
+	if windupWhen.Tofilemodel != nil {
+		// conditions = append(conditions, map[string]interface{}{"tofilemodel": nil})
+	}
+	// Test related
+	if windupWhen.Classificationexists != nil {
+		// conditions = append(conditions, map[string]interface{}{"classification-exists": nil})
+	}
+	// Test related
+	if windupWhen.Hintexists != nil {
+		// conditions = append(conditions, map[string]interface{}{"hint-exists": nil})
+	}
+	// Test related
+	if windupWhen.Lineitemexists != nil {
+		// conditions = append(conditions, map[string]interface{}{"lineitem-exists": nil})
+	}
+	// Test related
+	if windupWhen.Technologystatisticsexists != nil {
+		// conditions = append(conditions, map[string]interface{}{"technology-statistics-exists": nil})
+	}
+	// Test related
+	if windupWhen.Technologytagexists != nil {
+		// conditions = append(conditions, map[string]interface{}{"technology-tag-exists": nil})
+	}
+	return conditions
+}
+
+func convertWindupDependencyToAnalyzer(windupDependency windup.Dependency) map[string]interface{} {
+	name := strings.Replace(strings.Join([]string{windupDependency.GroupId, windupDependency.ArtifactId}, "."), "{*}", "*", -1)
+	dependency := map[string]interface{}{}
+	if strings.Contains(name, "*") {
+		dependency["nameregex"] = name
+	} else {
+		dependency["name"] = name
+	}
+
+	if windupDependency.FromVersion != "" {
+		dependency["lowerbound"] = windupDependency.FromVersion
+	}
+	if windupDependency.ToVersion != "" {
+		dependency["upperbound"] = windupDependency.ToVersion
+	}
+	if windupDependency.FromVersion == "" && windupDependency.ToVersion == "" {
+		dependency["lowerbound"] = "0.0.0"
+	}
+	return dependency
+}
+
+func convertWindupGraphQueryJarArchiveModel(gq windup.Graphquery) map[string]interface{} {
+	m := map[string]interface{}{
+		"nameregex": gq.Property.Value,
+	}
+	return m
+}
+
+// TODO handle perform fully
+func convertWindupPerformToAnalyzer(perform windup.Iteration, where map[string]string) map[string]interface{} {
+	ret := map[string]interface{}{}
+	tags := []string{}
+	if perform.Iteration != nil {
+		for _, it := range perform.Iteration {
+			converted := convertWindupPerformToAnalyzer(it, where)
+			for k, v := range converted {
+				ret[k] = v
+			}
+		}
+	}
+	if perform.Perform != nil {
+		converted := convertWindupPerformToAnalyzer(*perform.Perform, where)
+		for k, v := range converted {
+			ret[k] = v
+		}
+	}
+
+	if perform.Hint != nil {
+		if len(perform.Hint) != 1 {
+			// TODO
+			panic("More than one hint in a rule")
+			return nil
+		}
+		hint := perform.Hint[0]
+		if hint.Message != "" {
+			message := trimMessage(hint.Message)
+			// Handle some message replacement
+			message = strings.Replace(message, "{package}.{prefix}{type}", "{{name}}", -1)
+			message = strings.Replace(message, "{prefix}{type}", "{{type}}", -1)
+			ret["message"] = message
+		}
+	}
+	if perform.Technologyidentified != nil {
+		for _, ti := range perform.Technologyidentified {
+			for _, tag := range ti.Tag {
+				tags = append(tags, ti.Name+" - "+tag.Name)
+			}
+		}
+	}
+	if perform.Technologytag != nil {
+		for _, tag := range perform.Technologytag {
+			tags = append(tags, tag.Value)
+		}
+	}
+	if perform.Classification != nil {
+		for _, classification := range perform.Classification {
+			if classification.Tag != nil {
+				tags = append(tags, classification.Tag...)
+			}
+			if classification.Title != "" {
+				tags = append(tags, classification.Title)
+			}
+		}
+		// TODO perform.Classification.(Link|Effort|Categoryid|Of|Description|Quickfix|Issuedisplaymode)
+	}
+	if perform.Lineitem != nil {
+		if len(perform.Lineitem) != 1 {
+			// TODO
+			panic("More than one hint in a rule")
+			return nil
+		}
+		li := perform.Lineitem[0]
+		if li.Message != "" {
+			message := trimMessage(li.Message)
+			// Handle some message replacement
+			message = strings.Replace(message, "{package}.{prefix}{type}", "{{name}}", -1)
+			message = strings.Replace(message, "{prefix}{type}", "{{type}}", -1)
+			ret["message"] = message
+		}
+	}
+	if ret["tag"] != nil {
+		ret["tag"] = append(ret["tag"].([]string), tags...)
+	} else {
+		ret["tag"] = tags
+	}
+	return ret
+}
+
+func substituteWhere(where map[string]string, pattern string) string {
+	newString := pattern
+	for k, v := range where {
+		newString = strings.ReplaceAll(newString, "{"+k+"}", v)
+	}
+	return newString
+}
+
+func trimMessage(s string) string {
+	lines := strings.Split(s, "\n")
+	cleanLines := []string{}
+	for _, line := range lines {
+		cleaned := strings.Trim(line, "\n \t '")
+		if cleaned != "" {
+			cleanLines = append(cleanLines, cleaned)
+		}
+	}
+	return strings.Join(cleanLines, ". ")
+}
+
+func flattenWhere(wheres []windup.Where) map[string]string {
+	patterns := map[string]string{}
+	if wheres != nil && len(wheres) > 0 {
+		for _, where := range wheres {
+			// Where.Matches seems to always be length 1
+			patterns[where.Param] = where.Matches[0].Pattern
+		}
+	}
+	return patterns
+}
+
+func writeYAML(content interface{}, dest string) error {
+	file, err := os.OpenFile(dest, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0666)
+	if err != nil {
+		return err
+	}
+	defer file.Close()
+	enc := yaml.NewEncoder(file)
+
+	err = enc.Encode(content)
+	if err != nil {
+		return err
+	}
+	return nil
+}
